requirements.txt
python-magic-bin==0.4.14
pefile==2023.2.7
Pillow==10.1.0
requests==2.31.0
python-dotenv==1.0.0